fix update script

+++ update.bash
@@ -1,13 +1,12 @@
+if [ ! -f ".updated" ]; then
 
-# need to delete from replit pip packages
+  pip install --upgrade pip;
+  poetry install --no-root --no-dev --remove-untracked;
 
-pip install --upgrade pip;
-poetry install --no-root;
+  touch .updated;
 
-# download only if not found
-#python -m spacy info --silent $APP_SPACY_MODEL || python -m spacy download $APP_SPACY_MODEL;
-#python -m spacy info --exclude "spaCy version",Location,Platform,"Python version" | grep $APP_SPACY_MODEL || python -m spacy download $APP_SPACY_MODEL;
-#python -m spacy info | grep Pipelines | grep $APP_SPACY_MODEL || python -m spacy download $APP_SPACY_MODEL;
+  git add -A;
+  git rm .updated;
+  git diff-index --quiet HEAD || git commit -am 'install packages';
 
-git add -A;
-git diff-index --quiet HEAD || git commit -am 'install packages';
+fi
